requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file

blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
faiss-cpu==1.8.0.post1
filelock==3.15.4
flask==3.0.3
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==8.0.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80    
packaging==24.1
pillow==10.4.0
pip==20.2.3
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
setuptools==49.2.1
sympy==1.13.0
tbb==2021.13.0
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.4
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.19.2
